re-organize files and folders in build-system package
