+++ packages/hardhat-build-system/src/internal/builtin-tasks/utils/solidity-files-cache.ts
