+++ packages/hardhat-build-system/src/internal/solidity/compiler/solc-info.ts
