+++ packages/hardhat-build-system/src/internal/types/artifacts.ts
