+++ packages/hardhat-build-system/src/internal/utils/constants.ts
