+++ packages/hardhat-build-system/src/internal/utils/fs-utils.ts
@@ -1,7 +1,7 @@
 import fsPromises from "fs/promises";
 import path from "path";
 import fs from "fs";
-import { CustomError } from "./errors";
+import { CustomError } from "../errors/errors";
 
 /**
  * Returns an array of files (not dirs) that match a condition.
