+++ packages/hardhat-build-system/src/internal/utils/global-dir.ts
