+++ packages/hardhat-build-system/src/internal/utils/hash.ts
