+++ packages/hardhat-build-system/src/internal/utils/keccak.ts
