+++ packages/hardhat-build-system/src/internal/utils/package-info.ts
@@ -1,7 +1,7 @@
 import findup from "find-up";
 import fsExtra from "fs-extra";
 import path from "path";
-import { assertHardhatInvariant } from "./errors";
+import { assertHardhatInvariant } from "../errors/errors";
 
 export interface PackageJson {
   name: string;
@@ -17,7 +17,7 @@ export function getHardhatVersion(): string {
 
   assertHardhatInvariant(
     packageJsonPath !== null,
-    "There should be a package.json in hardhat-core's root directory"
+    "There should be a package.json in hardhat-core's root directory",
   );
 
   const packageJson = fsExtra.readJsonSync(packageJsonPath);
