+++ packages/hardhat-build-system/src/internal/utils/string.ts
